% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renal.R
\name{dosing_weight}
\alias{dosing_weight}
\title{Dosing weight for Cockcroft-Gault}
\usage{
dosing_weight(weight_kg, ibw_kg, obesity_ratio = 1.2, adj_factor = 0.4)
}
\arguments{
\item{weight_kg}{Actual body weight in kg.}

\item{ibw_kg}{Ideal body weight in kg.}

\item{obesity_ratio}{ABW/IBW ratio above which AdjBW is used.}

\item{adj_factor}{AdjBW correction factor.}
}
\value{
Dosing weight in kg. Vectorised.
}
\description{
If the patient is >20\% over IBW, ABW overestimates clearance, so AdjBW is used.
If they are underweight (ABW < IBW), ABW is used to avoid overdosing.
Otherwise IBW is used.
}
